# docs/swagger.yaml
basePath: /
definitions:
  github_com_Marwan051_final_project_backend_internal_service_agent.AgentRequest:
    properties:
      session_id:
        type: string
      user_query:
        type: string
    type: object
  github_com_Marwan051_final_project_backend_internal_service_agent.AgentResponse:
    properties:
      answer:
        type: string
      session_id:
        type: string
    type: object
  github_com_Marwan051_final_project_backend_internal_service_db_tools_proto.NearbyTripResult:
    properties:
      closest_stop_id:
        type: string
      closest_stop_lat:
        type: number
      closest_stop_lon:
        type: number
      closest_stop_name:
        type: string
      closest_stop_name_ar:
        type: string
      closest_stop_sequence:
        type: integer
      direction_id:
        type: integer
      distance_m:
        type: number
      route_id:
        type: string
      route_name:
        type: string
      route_name_ar:
        type: string
      route_short_name:
        type: string
      route_short_name_ar:
        type: string
      trip_headsign:
        type: string
      trip_headsign_ar:
        type: string
      trip_id:
        type: string
    type: object
  github_com_Marwan051_final_project_backend_internal_service_db_tools_proto.NearbyTripsRequest:
    properties:
      epsg:
        type: integer
      lat:
        type: number
      lon:
        type: number
      radius_m:
        type: number
      starts:
        type: boolean
    type: object
  github_com_Marwan051_final_project_backend_internal_service_db_tools_proto.NearbyTripsResponse:
    properties:
      count:
        type: integer
      epsg:
        type: integer
      lat:
        type: number
      lon:
        type: number
      radius_m:
        type: number
      starts:
        type: boolean
      trips:
        items:
          $ref: '#/definitions/github_com_Marwan051_final_project_backend_internal_service_db_tools_proto.NearbyTripResult'
        type: array
    type: object
  github_com_Marwan051_final_project_backend_internal_service_geocoding_proto.GeocodeRequest:
    properties:
      address:
        type: string
      bias:
        type: boolean
      language:
        type: string
      user_lat:
        description: 0 = not provided
        type: number
      user_lng:
        type: number
    type: object
  github_com_Marwan051_final_project_backend_internal_service_geocoding_proto.GeocodeResponse:
    properties:
      bias:
        type: boolean
      count:
        type: integer
      error:
        type: string
      language:
        type: string
      query:
        type: string
      results:
        items:
          $ref: '#/definitions/github_com_Marwan051_final_project_backend_internal_service_geocoding_proto.GeocodeResult'
        type: array
      success:
        type: boolean
    type: object
  github_com_Marwan051_final_project_backend_internal_service_geocoding_proto.GeocodeResult:
    properties:
      formatted_address:
        type: string
      latitude:
        type: number
      longitude:
        type: number
    type: object
  github_com_Marwan051_final_project_backend_internal_service_routing.AdminOperationResponse:
    properties:
      message:
        type: string
      status:
        type: string
      trips_reloaded:
        type: integer
    type: object
  github_com_Marwan051_final_project_backend_internal_service_routing.FilterBlock:
    properties:
      exclude:
        items:
          type: string
        type: array
      include:
        items:
          type: string
        type: array
      include_match:
        type: string
    type: object
  github_com_Marwan051_final_project_backend_internal_service_routing.Filters:
    properties:
      main_streets:
        $ref: '#/definitions/github_com_Marwan051_final_project_backend_internal_service_routing.FilterBlock'
      modes:
        $ref: '#/definitions/github_com_Marwan051_final_project_backend_internal_service_routing.FilterBlock'
    type: object
  github_com_Marwan051_final_project_backend_internal_service_routing.Journey:
    properties:
      id:
        type: integer
      labels:
        items:
          type: string
        type: array
      labels_ar:
        items:
          type: string
        type: array
      legs:
        items:
          $ref: '#/definitions/github_com_Marwan051_final_project_backend_internal_service_routing.Leg'
        type: array
      recommended_for:
        type: string
      summary:
        $ref: '#/definitions/github_com_Marwan051_final_project_backend_internal_service_routing.JourneySummary'
      text_summary:
        type: string
      text_summary_en:
        type: string
    type: object
  github_com_Marwan051_final_project_backend_internal_service_routing.JourneySummary:
    properties:
      cost:
        type: number
      main_streets_ar:
        items:
          type: string
        type: array
      main_streets_en:
        items:
          type: string
        type: array
      modes_ar:
        items:
          type: string
        type: array
      modes_en:
        items:
          type: string
        type: array
      total_distance_meters:
        type: integer
      total_time_minutes:
        type: integer
      transfers:
        type: integer
      transit_distance_meters:
        type: integer
      walking_distance_meters:
        type: integer
    type: object
  github_com_Marwan051_final_project_backend_internal_service_routing.Leg:
    properties:
      distance_meters:
        type: integer
      duration_minutes:
        type: integer
      end_stop_id:
        type: string
      fare:
        type: number
      from_stop:
        $ref: '#/definitions/github_com_Marwan051_final_project_backend_internal_service_routing.StopInfo'
      from_trip_id:
        type: string
      from_trip_name:
        type: string
      from_trip_name_ar:
        type: string
      headsign:
        type: string
      headsign_ar:
        type: string
      mode_ar:
        type: string
      mode_en:
        type: string
      polyline:
        type: string
      route_short_name:
        type: string
      route_short_name_ar:
        type: string
      to_stop:
        $ref: '#/definitions/github_com_Marwan051_final_project_backend_internal_service_routing.StopInfo'
      to_trip_id:
        type: string
      to_trip_name:
        type: string
      to_trip_name_ar:
        type: string
      trip_id:
        type: string
      trip_ids:
        items:
          type: string
        type: array
      type:
        type: string
      walking_distance_meters:
        type: integer
    type: object
  github_com_Marwan051_final_project_backend_internal_service_routing.RouteRequest:
    properties:
      end_lat:
        type: number
      end_lon:
        type: number
      filters:
        $ref: '#/definitions/github_com_Marwan051_final_project_backend_internal_service_routing.Filters'
      max_transfers:
        type: integer
      priority:
        type: string
      start_lat:
        type: number
      start_lon:
        type: number
      top_k:
        type: integer
      walking_cutoff:
        type: integer
      weights:
        additionalProperties:
          format: float64
          type: number
        type: object
    type: object
  github_com_Marwan051_final_project_backend_internal_service_routing.RouteResponse:
    properties:
      end_trips_found:
        type: integer
      error:
        type: string
      geometry_encoding:
        type: string
      journeys:
        items:
          $ref: '#/definitions/github_com_Marwan051_final_project_backend_internal_service_routing.Journey'
        type: array
      num_journeys:
        type: integer
      selected_priority:
        type: string
      start_trips_found:
        type: integer
      total_after_dedup:
        type: integer
      total_routes_found:
        type: integer
      weights_used:
        additionalProperties:
          format: float64
          type: number
        type: object
    type: object
  github_com_Marwan051_final_project_backend_internal_service_routing.StopInfo:
    properties:
      coord:
        items:
          type: number
        type: array
      name:
        type: string
      name_ar:
        type: string
      stop_id:
        type: string
    type: object
  github_com_Marwan051_final_project_backend_internal_service_traffic_updater_proto.StatusResponse:
    properties:
      is_running:
        type: boolean
      last_update:
        type: string
      status:
        type: string
      trips_in_data:
        type: integer
    type: object
  github_com_Marwan051_final_project_backend_internal_service_traffic_updater_proto.StreetInfo:
    properties:
      aliases:
        items:
          type: string
        type: array
      name:
        type: string
      name_ar:
        type: string
      segments:
        type: integer
      total_length_km:
        type: number
    type: object
  github_com_Marwan051_final_project_backend_internal_service_traffic_updater_proto.StreetListResponse:
    properties:
      count:
        type: integer
      streets:
        items:
          $ref: '#/definitions/github_com_Marwan051_final_project_backend_internal_service_traffic_updater_proto.StreetInfo'
        type: array
    type: object
  github_com_Marwan051_final_project_backend_internal_service_traffic_updater_proto.StreetTrafficLeg:
    properties:
      distance_m:
        type: integer
      distance_text:
        type: string
      duration_seconds:
        type: integer
      duration_text:
        type: string
    type: object
  github_com_Marwan051_final_project_backend_internal_service_traffic_updater_proto.StreetTrafficRequest:
    properties:
      language:
        type: string
      max_waypoints:
        type: integer
      name:
        type: string
    type: object
  github_com_Marwan051_final_project_backend_internal_service_traffic_updater_proto.StreetTrafficResponse:
    properties:
      error:
        type: string
      legs:
        items:
          $ref: '#/definitions/github_com_Marwan051_final_project_backend_internal_service_traffic_updater_proto.StreetTrafficLeg'
        type: array
      routes:
        items:
          $ref: '#/definitions/github_com_Marwan051_final_project_backend_internal_service_traffic_updater_proto.StreetTrafficRoute'
        type: array
      segments:
        type: integer
      street:
        type: string
      street_ar:
        type: string
      total_distance_km:
        type: number
      total_duration_min:
        type: number
      waypoints_used:
        type: integer
    type: object
  github_com_Marwan051_final_project_backend_internal_service_traffic_updater_proto.StreetTrafficRoute:
    properties:
      distance_m:
        type: integer
      distance_text:
        type: string
      duration_seconds:
        type: integer
      duration_text:
        type: string
      label:
        type: string
    type: object
  github_com_Marwan051_final_project_backend_internal_service_traffic_updater_proto.TriggerRequest:
    properties:
      notify_routing_api:
        type: boolean
      trip_ids:
        description: empty = all trips
        items:
          type: string
        type: array
    type: object
  github_com_Marwan051_final_project_backend_internal_service_traffic_updater_proto.UpdateResponse:
    properties:
      message:
        type: string
      status:
        type: string
      trips_failed:
        type: integer
      trips_updated:
        type: integer
    type: object
  github_com_Marwan051_final_project_backend_internal_service_traffic_updater_proto.UpdateTripRequest:
    properties:
      notify_routing_api:
        type: boolean
      trip_id:
        type: string
    type: object
  internal_api_v1.HealthResponse:
    properties:
      status:
        type: string
      timestamp:
        type: string
    type: object
info:
  contact: {}
  description: This is the API server for the Routing App Backend.
  title: Routing App Backend API
  version: "1.0"
paths:
  /api/v1/agent/query:
    post:
      consumes:
      - application/json
      description: Sends a message to the agent and returns the reply plus session
        id
      parameters:
      - description: Agent Query Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_Marwan051_final_project_backend_internal_service_agent.AgentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Marwan051_final_project_backend_internal_service_agent.AgentResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Agent Query
      tags:
      - agent
  /api/v1/geocode:
    post:
      consumes:
      - application/json
      description: Converts a text address into latitude and longitude coordinates
      parameters:
      - description: Geocode Request (Address)
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_Marwan051_final_project_backend_internal_service_geocoding_proto.GeocodeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Marwan051_final_project_backend_internal_service_geocoding_proto.GeocodeResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Geocode Address
      tags:
      - geocoding
  /api/v1/nearby-trips:
    post:
      consumes:
      - application/json
      description: Finds trips near a given location within a specified radius
      parameters:
      - description: Nearby Trips Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_Marwan051_final_project_backend_internal_service_db_tools_proto.NearbyTripsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Marwan051_final_project_backend_internal_service_db_tools_proto.NearbyTripsResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get Nearby Trips
      tags:
      - db_tools
  /api/v1/route:
    post:
      consumes:
      - application/json
      description: Calculates the best route from start coordinates to end coordinates
      parameters:
      - description: Route Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_Marwan051_final_project_backend_internal_service_routing.RouteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Marwan051_final_project_backend_internal_service_routing.RouteResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Find Route
      tags:
      - routing
  /api/v1/routing/rebuild-network:
    post:
      consumes:
      - application/json
      description: Admin only. Rebuilds the routing network from source data.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Marwan051_final_project_backend_internal_service_routing.AdminOperationResponse'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Rebuild Routing Network
      tags:
      - routing
  /api/v1/routing/reload-prefix-times:
    post:
      consumes:
      - application/json
      description: Admin only. Reloads prefix times for the routing graph.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Marwan051_final_project_backend_internal_service_routing.AdminOperationResponse'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Reload Prefix Times
      tags:
      - routing
  /api/v1/traffic/status:
    get:
      description: Retrieves the current traffic update status
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Marwan051_final_project_backend_internal_service_traffic_updater_proto.StatusResponse'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get Traffic Status
      tags:
      - traffic
  /api/v1/traffic/street:
    post:
      consumes:
      - application/json
      description: Fetches the current calculated traffic load on a specific street
      parameters:
      - description: Get Street Traffic Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_Marwan051_final_project_backend_internal_service_traffic_updater_proto.StreetTrafficRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Marwan051_final_project_backend_internal_service_traffic_updater_proto.StreetTrafficResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get Street Traffic
      tags:
      - traffic
  /api/v1/traffic/streets:
    get:
      description: Returns a list of all streets actively tracked and their traffic
        levels
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Marwan051_final_project_backend_internal_service_traffic_updater_proto.StreetListResponse'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: List All Streets Traffic
      tags:
      - traffic
  /api/v1/traffic/trigger:
    post:
      consumes:
      - application/json
      description: Triggers a background traffic update process
      parameters:
      - description: Trigger Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_Marwan051_final_project_backend_internal_service_traffic_updater_proto.TriggerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Marwan051_final_project_backend_internal_service_traffic_updater_proto.UpdateResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Trigger Traffic Update
      tags:
      - traffic
  /api/v1/traffic/update-trip:
    post:
      consumes:
      - application/json
      description: Submits trip information to affect current traffic modelling
      parameters:
      - description: Update Trip Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_Marwan051_final_project_backend_internal_service_traffic_updater_proto.UpdateTripRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Marwan051_final_project_backend_internal_service_traffic_updater_proto.UpdateResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Update Trip Traffic
      tags:
      - traffic
  /health:
    get:
      consumes:
      - application/json
      description: Returns the health status of the backend service
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api_v1.HealthResponse'
      summary: Health Check
      tags:
      - system
swagger: "2.0"