# internal/api/v1/routes.go
package v1

import (
	"net/http"
	"time"

	"github.com/Marwan051/final_project_backend/internal/api/v1/handlers"
	authctx "github.com/Marwan051/final_project_backend/internal/authctx"
	agent_service "github.com/Marwan051/final_project_backend/internal/service/agent"
	"github.com/Marwan051/final_project_backend/internal/service/db_tools"
	"github.com/Marwan051/final_project_backend/internal/service/geocoding"
	route_service "github.com/Marwan051/final_project_backend/internal/service/routing"
	"github.com/Marwan051/final_project_backend/internal/service/traffic_updater"
	"github.com/Marwan051/final_project_backend/internal/utils"
)

// NewRouter returns a new router with all v1 API routes
func NewRouter(
	routingService route_service.Router,
	agentService agent_service.Agent,
	dbToolsService db_tools.DbTools,
	geocodingService geocoding.Geocoding,
	trafficService traffic_updater.TrafficUpdater,
) *http.ServeMux {
	mux := http.NewServeMux()

	routingHandler := handlers.NewRoutingHandler(routingService)
	agentHandler := handlers.NewAgentHandler(agentService)
	dbToolsHandler := handlers.NewDbToolsHandler(dbToolsService)
	geocodingHandler := handlers.NewGeocodingHandler(geocodingService)
	trafficHandler := handlers.NewTrafficHandler(trafficService)

	mux.HandleFunc("GET /health", HealthHandler)

	// Routing endpoint
	mux.HandleFunc("POST /route", routingHandler.FindRoute)
	mux.Handle("POST /routing/reload-prefix-times", requireAdmin(http.HandlerFunc(routingHandler.ReloadPrefixTimes)))
	mux.Handle("POST /routing/rebuild-network", requireAdmin(http.HandlerFunc(routingHandler.RebuildNetwork)))

	// Agent endpoint
	mux.HandleFunc("POST /agent/query", agentHandler.Query)

	// Geocoding endpoints
	mux.HandleFunc("POST /geocode", geocodingHandler.Geocode)

	// DB Tools endpoints
	mux.HandleFunc("POST /nearby-trips", dbToolsHandler.NearbyTrips)

	// Traffic Updater endpoints
	mux.HandleFunc("POST /traffic/trigger", trafficHandler.TriggerUpdate)
	mux.HandleFunc("GET /traffic/status", trafficHandler.GetStatus)
	mux.HandleFunc("POST /traffic/update-trip", trafficHandler.UpdateTrip)
	mux.HandleFunc("POST /traffic/street", trafficHandler.StreetTraffic)
	mux.HandleFunc("GET /traffic/streets", trafficHandler.ListStreets)

	return mux
}

type HealthResponse struct {
	Status    string `json:"status"`
	Timestamp string `json:"timestamp"`
}

// HealthHandler returns the health status of the service
// @Summary      Health Check
// @Description  Returns the health status of the backend service
// @Tags         system
// @Accept       json
// @Produce      json
// @Success      200  {object}  HealthResponse
// @Router       /health [get]
func HealthHandler(w http.ResponseWriter, r *http.Request) {
	response := HealthResponse{
		Status:    "ok",
		Timestamp: time.Now().UTC().Format(time.RFC3339),
	}

	utils.WriteJSONResponse(w, http.StatusOK, response)
}

func requireAdmin(next http.Handler) http.Handler {
	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		if utils.Cfg.ENV == "dev" && utils.Cfg.DisableAuth {
			http.Error(w, "forbidden", http.StatusForbidden)
			return
		}
		claims, ok := authctx.GetUserClaims(r.Context())
		if !ok {
			http.Error(w, "forbidden", http.StatusForbidden)
			return
		}
		if role, ok := claims["role"].(string); ok && role == "admin" {
			next.ServeHTTP(w, r)
			return
		}
		if role, ok := claims["app_metadata"].(map[string]any); ok {
			if v, ok := role["role"].(string); ok && v == "admin" {
				next.ServeHTTP(w, r)
				return
			}
			if v, ok := role["admin"].(bool); ok && v {
				next.ServeHTTP(w, r)
				return
			}
		}
		if admin, ok := claims["admin"].(bool); ok && admin {
			next.ServeHTTP(w, r)
			return
		}
		http.Error(w, "forbidden", http.StatusForbidden)
	})
}